Added pagination to front page
Without pagination, the initial load is several mbs of profile images.
At this time, this is a non-negligible amount of data (450+ profiles).
This could be further improved by controlling the manner in which
profile pictures are provided, so that image optimization can be
implemented.

Added a pagination component that can be configured to load a number of
profiles. It provides a "load more" button to continue to load profiles
incrementally.

Sketch particles plugin was disabled, as it is an unmaintained plugin,
does not provide any documentation on its use, and effectively disables
all click events to areas where the plugin is active.