build: switched to .prettierrcjson with default config from documentation (#2939)
It seems like using the `.prettierrc` file doesn't work reliably, as I've seen a pull request that changed files but didn't trigger
the lint-staged script (with prettier). When I manually run the prettier script, it also does not consider the `semi:false`.

Using `.prettierrc.json` seems to work, and it also has the advantage of being a `.json` file, so it can be formatted and displayed nicely.